requirements.txt
annotated-types==0.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
et_xmlfile==2.0.0
idna==3.10
inflection==0.5.1
numpy==2.1.3
openpyxl==3.1.5
pandas==2.2.3
pyairtable==2.3.6
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
XlsxWriter==3.2.0
